requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0
pypdf>=4.0.0
streamlit>=1.40.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
schedule>=1.2.0
